added move endpoint for shareFood

+++ src/main/java/no/ntnu/stud/idatt2106/backend/model/response/FoodSummaryResponse.java
@@ -8,5 +8,5 @@ import lombok.Data;
 @Data
 public class FoodSummaryResponse {
   private int typeId;
-  private int totalAmount;
+  private float totalAmount;
 }
